parkaran-helper/requirements.txt
flask==3.1.0
chromadb>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
ollama>=0.4.0
requests>=2.31.0
